stop being a big stupid..
I was brand new to GnuPG/PGP, and cryptography in general.
I wanted to used a super sophisticated system and have this super cool and secure backend for all my shenanigans.
But really it was all too complex and I ended up just procrastinating everything and not getting any work done.

Maturing is realizing you really don't need much, and no one is going to be targeting my for a long while.